ron_skills/sql_query_optimizer/api/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
anthropic==0.86.0
pydantic==2.12.5
python-dotenv==1.2.1
stripe==15.0.0
slowapi==0.1.9
